Cross_Border_Project/requirements.txt
# HTTP and API clients
httpx>=0.27.0
anthropic>=0.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.7.0

# Logging
structlog>=24.1.0
colorama>=0.4.6

# Scheduling
croniter>=2.0.0

# Data processing
pandas>=2.2.0

# Template rendering
Jinja2>=3.1.3

# AI token counting
tiktoken>=0.6.0

# Ad platform SDKs
facebook-business>=19.0.0
google-ads>=23.0.0

# Reporting output
rich>=13.0.0

# Utilities
diskcache>=5.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
